api/requirements.txt
alembic==0.9.3
aniso8601==1.2.1
argon2-cffi==16.3.0
blinker==1.4
cffi==1.10.0
click==6.7
ecdsa==0.13
Flask==0.12.2
Flask-Mail==0.9.1
Flask-Migrate==2.0.4
Flask-RESTful==0.3.6
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
future==0.16.0
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.7
MarkupSafe==1.0
passlib==1.7.1
pycparser==2.18
pycrypto==2.6.1
python-dateutil==2.6.1
python-editor==1.0.3
python-jose==1.3.2
pytz==2017.2
six==1.10.0
SQLAlchemy==1.1.11
Werkzeug==0.12.2
